Reorganize newtmgr commands
* Move commands to cli package in newtmgr/cli
* Place each command in its own file in the cli directory
* Move cli/cp.go => config/connprofile.go
  This file is responsible for managing config storage and doesn't
  implement commands
