+++ newtmgr/cli/echo.go
@@ -1,0 +1,69 @@
+package cli
+
+import (
+	"fmt"
+
+	"git-wip-us.apache.org/repos/asf/incubator-mynewt-newt/newtmgr/config"
+	"git-wip-us.apache.org/repos/asf/incubator-mynewt-newt/newtmgr/protocol"
+	"git-wip-us.apache.org/repos/asf/incubator-mynewt-newt/newtmgr/transport"
+	"github.com/spf13/cobra"
+)
+
+func echoRunCmd(cmd *cobra.Command, args []string) {
+	cpm, err := config.NewConnProfileMgr()
+	if err != nil {
+		nmUsage(cmd, err)
+	}
+
+	profile, err := cpm.GetConnProfile(ConnProfileName)
+	if err != nil {
+		nmUsage(cmd, err)
+	}
+
+	conn, err := transport.NewConn(profile)
+	if err != nil {
+		nmUsage(cmd, err)
+	}
+
+	runner, err := protocol.NewCmdRunner(conn)
+	if err != nil {
+		nmUsage(cmd, err)
+	}
+
+	echo, err := protocol.NewEcho()
+	if err != nil {
+		nmUsage(cmd, err)
+	}
+
+	echo.Message = args[0]
+
+	nmr, err := echo.EncodeWriteRequest()
+	if err != nil {
+		nmUsage(cmd, err)
+	}
+
+	if err := runner.WriteReq(nmr); err != nil {
+		nmUsage(cmd, err)
+	}
+
+	rsp, err := runner.ReadResp()
+	if err != nil {
+		nmUsage(cmd, err)
+	}
+
+	ersp, err := protocol.DecodeEchoResponse(rsp.Data)
+	if err != nil {
+		nmUsage(cmd, err)
+	}
+	fmt.Println(ersp.Message)
+}
+
+func echoCmd() *cobra.Command {
+	echoCmd := &cobra.Command{
+		Use:   "echo",
+		Short: "Send data to remote endpoint using newtmgr, and receive data back",
+		Run:   echoRunCmd,
+	}
+
+	return echoCmd
+}
