+++ newtmgr/config/connprofile.go
@@ -1,23 +1,4 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package cli
+package config
 
 import (
 	"log"
@@ -26,7 +7,7 @@ import (
 	"github.com/mitchellh/go-homedir"
 )
 
-type CpMgr struct {
+type ConnProfileMgr struct {
 	cDb *util.CfgDb
 }
 
@@ -42,8 +23,8 @@ type ConnProfile struct {
 	MyConnString string
 }
 
-func NewCpMgr() (*CpMgr, error) {
-	cpm := &CpMgr{}
+func NewConnProfileMgr() (*ConnProfileMgr, error) {
+	cpm := &ConnProfileMgr{}
 
 	if err := cpm.Init(); err != nil {
 		return nil, err
@@ -52,7 +33,7 @@ func NewCpMgr() (*CpMgr, error) {
 	return cpm, nil
 }
 
-func (cpm *CpMgr) Init() error {
+func (cpm *ConnProfileMgr) Init() error {
 	var err error
 
 	dir, err := homedir.Dir()
@@ -68,7 +49,7 @@ func (cpm *CpMgr) Init() error {
 	return nil
 }
 
-func (cpm *CpMgr) GetConnProfileList() ([]*ConnProfile, error) {
+func (cpm *ConnProfileMgr) GetConnProfileList() ([]*ConnProfile, error) {
 	log.Printf("[DEBUG] Getting list of connection profiles")
 	cpMap, err := cpm.cDb.GetSect("conn_profile_list")
 	if err != nil {
@@ -89,7 +70,7 @@ func (cpm *CpMgr) GetConnProfileList() ([]*ConnProfile, error) {
 	return cpList, nil
 }
 
-func (cpm *CpMgr) DeleteConnProfile(name string) error {
+func (cpm *ConnProfileMgr) DeleteConnProfile(name string) error {
 	if err := cpm.cDb.DeleteSect("_conn_profile_" + name); err != nil {
 		return err
 	}
@@ -101,7 +82,7 @@ func (cpm *CpMgr) DeleteConnProfile(name string) error {
 	return nil
 }
 
-func (cpm *CpMgr) AddConnProfile(cp *ConnProfile) error {
+func (cpm *ConnProfileMgr) AddConnProfile(cp *ConnProfile) error {
 	sect := "_conn_profile_" + cp.Name()
 	cDb := cpm.cDb
 
@@ -116,7 +97,7 @@ func (cpm *CpMgr) AddConnProfile(cp *ConnProfile) error {
 	return nil
 }
 
-func (cpm *CpMgr) GetConnProfile(pName string) (*ConnProfile, error) {
+func (cpm *ConnProfileMgr) GetConnProfile(pName string) (*ConnProfile, error) {
 	// Each section is a connection profile, key values are the contents
 	// of that section.
 	if pName == "" {
